Add creation of bin dir to Makefile
Contrary to SVN Git doesn't support empty directories. We used to have
an empty bin/ directory where the binaries would end up when running
make.
This change adds a few rules to the Makefile to set the creation of the
bin/ dir as an order-only prerequisite for the binary files. See the
Make manual for details:
http://www.gnu.org/software/make/manual/make.html#index-prerequisite-types-140

Also added the bin/ directory to the .gitignore file.